Re-arrange utils package.
