+++ utils/log/log_unsupported.go
@@ -1,6 +1,6 @@
 // +build !linux
 
-package utils
+package log
 
 // MaybeSetupLoggerIfOnJournaldAvailable sets up journald logging if the system
 func MaybeSetupLoggerIfOnJournaldAvailable() {
