+++ utils/netns/nsdialer_linux.go
@@ -1,19 +1,19 @@
 // +build linux
 
-package utils
+package netns
 
 import (
-	"fmt"
+	"context"
 	"net"
 	"runtime"
 
-	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
+	"go.uber.org/multierr"
 
 	"github.com/vishvananda/netns"
 )
 
-func GetNsListener(netNsPath string, port int) (net.Listener, error) {
+func (n *NsDialer) DialContext(ctx context.Context, network, address string) (net.Conn, error) {
 	runtime.LockOSThread()
 	defer func() {
 		runtime.UnlockOSThread()
@@ -27,7 +27,7 @@ func GetNsListener(netNsPath string, port int) (net.Listener, error) {
 		_ = origNs.Close()
 	}()
 
-	dialNs, err := netns.GetFromPath(netNsPath)
+	dialNs, err := netns.GetFromPath(n.netNsPath)
 	if err != nil {
 		return nil, errors.Wrap(err, "Unable to open namespace")
 	}
@@ -35,21 +35,17 @@ func GetNsListener(netNsPath string, port int) (net.Listener, error) {
 		_ = dialNs.Close()
 	}()
 
-	var errs *multierror.Error
-
 	err = netns.Set(dialNs)
 	if err != nil {
 		return nil, errors.Wrap(err, "Unable to enter namespace")
 	}
 
-	listener, err := net.Listen("tcp", fmt.Sprintf(":%d", port))
-	errs = multierror.Append(errs, err)
+	conn, err := n.dialer.DialContext(ctx, network, address)
 
-	err = netns.Set(origNs)
-	if err != nil {
-		err = errors.Wrap(err, "Unable to restore namespace")
+	errA := netns.Set(origNs)
+	if errA != nil {
+		errA = errors.Wrap(err, "Unable to restore namespace")
 	}
-	errs = multierror.Append(errs, err)
 
-	return listener, errs.ErrorOrNil()
+	return conn, multierr.Combine(err, errA)
 }
